fix: Fix bug when using a namespace starting with App (example: AppleController.eat)
Hi, I had this error when I added a controller starting with App, I found the bug in the

Helpers.makeNameSpace() method. I just replaced toPath.startsWith(appNameSpace) to

toPath.startsWith(`/`) to make sure it check the slash after the namespace. I also

added a test to verify any regression for this bug :-)